Add GairqGeo class
This class provides the features of requesting geolocalized feed
to the API server.

The object can be instantiated by two ways of ip base and
latitude/longitude. The former way is by the current ip of the host
and the latter one is by the information as it is, latitude and
longitude. However, when it succeeds in instantiating the class,
developers can request the geo location based information with sync and
async methods.

Note that async method fits for GUI programming.